+++ organize/filters/created.py
@@ -1,15 +1,13 @@
 import subprocess
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from typing import Union
 
 from fs.base import FS
-from schema import Optional, Or
 
-from .filter import Filter, FilterResult
-from .utils import age_condition_applies
+from ._timefilter import TimeFilter
 
 
-class Created(Filter):
+class Created(TimeFilter):
     """Matches files / folders by created date
 
     Args:
@@ -30,70 +28,16 @@ class Created(Filter):
     """
 
     name = "created"
-    schema_support_instance_without_args = True
-    arg_schema = {
-        Optional("years"): int,
-        Optional("months"): int,
-        Optional("weeks"): int,
-        Optional("days"): int,
-        Optional("hours"): int,
-        Optional("minutes"): int,
-        Optional("seconds"): int,
-        Optional("mode"): Or("older", "newer"),
-    }
 
-    def __init__(
-        self,
-        years=0,
-        months=0,
-        weeks=0,
-        days=0,
-        hours=0,
-        minutes=0,
-        seconds=0,
-        mode="older",
-    ):
-        self.age = timedelta(
-            weeks=52 * years + 4 * months + weeks,  # quick and a bit dirty
-            days=days,
-            hours=hours,
-            minutes=minutes,
-            seconds=seconds,
-        )
-        self.mode = mode.strip().lower()
-        if self.mode not in ("older", "newer"):
-            raise ValueError("Unknown option for 'mode': must be 'older' or 'newer'.")
-
-    def matches_created_time(self, created: Union[None, datetime]):
-        match = True
-        if self.age.total_seconds():
-            if not created:
-                match = False
-            else:
-                match = age_condition_applies(
-                    dt=created,
-                    age=self.age,
-                    mode=self.mode,
-                    reference=datetime.now(),
-                )
-        return match
-
-    def pipeline(self, args: dict) -> FilterResult:
+    def get_datetime(self, args) -> Union[None, datetime]:
         fs = args["fs"]  # type: FS
         fs_path = args["fs_path"]
-
         created = fs.getinfo(fs_path, namespaces=["details"]).created
         if not created:
             # We're probably on Linux. No easy way to get creation dates here,
             # so we try to use the stat utility.
             created = self.fallback_method(fs, fs_path)
-
-        created = created.astimezone()
-        match = self.matches_created_time(created)
-        return FilterResult(
-            matches=match,
-            updates={self.get_name(): created},
-        )
+        return created
 
     def fallback_method(self, fs, fs_path):
         if fs.hassyspath(fs_path):