+++ organize/filters/date_added.py
@@ -1,3 +1,5 @@
+import subprocess
+import sys
 from datetime import datetime
 from typing import Union
 
@@ -6,9 +8,11 @@ from fs.base import FS
 from ._timefilter import TimeFilter
 
 
-class LastModified(TimeFilter):
+class DateAdded(TimeFilter):
 
-    """Matches files by last modified date
+    """Matches files by the time the file was added to a folder.
+
+    **`date_added` is only available on macOS!**
 
     Args:
         years (int): specify number of years
@@ -24,19 +28,29 @@ class LastModified(TimeFilter):
             time. (default = 'older')
 
     Returns:
-        {lastmodified}: The datetime the files / folders was lastmodified.
+        {date_added}: The datetime the files / folders were added.
     """
 
-    name = "lastmodified"
+    name = "date_added"
 
     def get_datetime(self, args) -> Union[None, datetime]:
+        if sys.platform != "darwin":
+            raise EnvironmentError("date_added is only available on macOS")
+
         fs = args["fs"]  # type: FS
         fs_path = args["fs_path"]
-        modified = fs.getmodified(fs_path)
-        return modified
+
+        out = subprocess.run(
+            ["mdls", "-name", "kMDItemDateAdded", "-raw", fs.getsyspath(fs_path)],
+            capture_output=True,
+            encoding="utf-8",
+            check=True,
+        ).stdout
+        dt = datetime.strptime(out, "%Y-%m-%d %H:%M:%S %z")
+        return dt
 
     def __str__(self):
-        return "[LastModified] All files / folders last modified %s than %s" % (
+        return "[DateAdded] All files / folders added %s than %s" % (
             self._mode,
             self.timedelta,
         )